+++ src/components/home/SecondSection.tsx
@@ -1,6 +1,5 @@
 'use client';
 
-import { useState, useEffect } from 'react';
 import Image from 'next/image';
 import Link from 'next/link';
 import { cn } from '@/lib/utils';
@@ -11,25 +10,7 @@ import { Button } from '@/components/ui';
 import { SectionHeader } from '@/components/home';
 
 export const SecondSection = () => {
-  const [isMd, setIsMd] = useState(false);
-
   const data = menuData.find((data) => data.title === '제품소개');
-
-  useEffect(() => {
-    const mediaQuery = window.matchMedia('(min-width: 768px)');
-
-    const handleChange = (event: MediaQueryListEvent) => {
-      setIsMd(event.matches);
-    };
-
-    // 초기값 설정
-    setIsMd(mediaQuery.matches);
-
-    // 사이즈 변경 감지
-    mediaQuery.addEventListener('change', handleChange);
-
-    return () => mediaQuery.removeEventListener('change', handleChange);
-  }, []);
 
   if (!data || !data.subMenu) {
     return (
@@ -58,8 +39,6 @@ export const SecondSection = () => {
         {data.subMenu.map((item, index) => (
           <InView key={item.label} threshold={0.2}>
             {({ ref, inView }) => {
-              const delay = isMd ? `${index * 200}ms` : '0ms';
-
               return (
                 <Link
                   href={item.url}
@@ -71,12 +50,9 @@ export const SecondSection = () => {
                       ? 'opacity-100 translate-y-0 md:translate-y-0 md:translate-x-0'
                       : 'opacity-0 translate-y-10 md:translate-y-0 md:translate-x-10'
                   )}
-                  style={{
-                    transitionDelay: delay,
-                  }}
                 >
                   {/* 이미지 */}
-                  <div className='flex items-center justify-center w-full h-[30rem] md:h-[380px] overflow-hidden'>
+                  <div className='flex items-center justify-center w-full h-[30rem] md:h-[380px] lg:h-[480px] overflow-hidden'>
                     <Image
                       src={item.img!}
                       alt={item.label}
